refactor: setElementAttributes 함수 공통화
